embedding_analysis/requirements.txt
jax
jaxlib
numpy
matplotlib
ml_collections
scipy
scikit-image
scikit-learn
tqdm
umap-learn
hdbscan